A bunch of wording changes, including merging the about and home pages.
